Return file snapshot on upload (#88)
#### What's this PR do?

The promise from 'uploadFile' now returns an object of the file data is decorated with a `pushKey` property that is the key generated by the call to `firebase.push`.  This change will give the caller the ability to access the urls, name and key for additional processing.

#### What are the important parts of the code?
a single `.then` with a spread

#### How should this be tested by the reviewer?
Ensure the caller of this function has access to the fileData in the returned promise.

#### Is any other information necessary to understand this?
This change is proposed as the caller may need to dispatch additional actions for use with redux/redux-observable that contain the information contained in the fileData.