fix: unwatch events with same parameters as original watch (#151)
unWatchEvent was not passing the correct queryId or path param (i.e. matching params passed to watchEvent), causing internal firebase._.watchers to continuously grow due to unsetWatcher being unable to resolve the watcher. This manifests as open firebase subscriptions continuing to grow, even when then the component that issued the query that opened the subscription is unmounted.

PR fixes the passed params so that unsetWatcher is able to clean up the watchers on unmount. Best test case for this is probably to compare that firebase._.watchers map after component mount and component unmount, and compare this PR against behavior on master.